% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplespectra}
\alias{examplespectra}
\title{Example spectra used by Spectran}
\format{
\code{examplespectra} A list of two elements, each containing a table of spectral data
\describe{
\item{Measurement}{Spectral data measured by a spectroradiometer. The unit of the wavelength (Wellenlaenge) is nm, the unit of the spectral data is W/m^2.}
\item{CIE}{Spectral data from the CIE describing Standard Illuminants. The unit of the wavelength (Wellenlaenge) is nm, the unit of the spectral data is W/m^2.}
}
}
\source{
\url{https://cie.co.at/data-tables}
}
\usage{
examplespectra
}
\description{
Contains Measurement and CIE Standard Illuminant Spectra. Additional data to the spectra can be found in the \link{examplespectra_descriptor} dataset.
}
\keyword{datasets}
