% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectral_analysis.R
\name{Spectran}
\alias{Spectran}
\title{Spectran}
\usage{
Spectran(lang_setting = "English", ...)
}
\arguments{
\item{lang_setting}{A language for the application. Currently \strong{Deutsch} for German and \strong{English} (default) are implemented.}

\item{...}{Any other settings that get passed to shinyApp}
}
\value{
Open a viewer with the shiny app
}
\description{
Spectran
}
