% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ColorP}
\alias{ColorP}
\title{Hex Color pallets for Spectran}
\format{
\code{ColorP} A list containing 3 elements, each a named vector of colors
\describe{
\item{regenbogen}{401 colors along the spectrum that make a nice, yet nonaccurate representation of the visible spectrum. Created by the \code{\link[grDevices:palettes]{grDevices::rainbow()}} function.}
\item{Lang}{401 colors along the spectrum that make a nice, yet nonaccurate representation of the visible spectrum. Created from a picture provided by \verb{Dieter Lang}.}
\item{Color_Rendering}{14 colors that represent the reference colors used to determine the color rendering index.}
\item{Alpha}{A list. \code{names} contains the names of the 5 human receptor types. \code{adjectives} contains these names in their adjective form for every languages included in Spectran. \code{descriptions} contains the descriptions used for tables and plotting. \code{abb} contains the abbreviations (if any) used for the equivalent daylight illuminance for that receptor type.}
\item{Vlambda}{A character scalar containing the properly escaped version of V(lambda), used e.g. in Plots}
\item{Alpha.ico}{A character scalar containing the properly escaped version of the sign \code{alpha}, used e.g. in Plots}
\item{Plot}{A table of Peak Wavelength (nm) values, Names, Abbreviations, and colors for the 5 human receptor types and V(lambda) from 1924. This table is mainly used for plotting.}

}
}
\source{
\code{\link[grDevices:palettes]{grDevices::rainbow()}}
}
\usage{
ColorP
}
\description{
Contains named vectors of Hex color codes for Spectran's color pallets.
}
\keyword{datasets}
