% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectran-package.R
\docType{package}
\name{Spectran-package}
\alias{Spectran-package}
\alias{_PACKAGE}
\alias{testduplicate-package}
\title{Spectran: Visual and Non-Visual Spectral Analysis of Light}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Johannes Zauner \email{spectran@litg.de} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
